+++ src/frontend/apps/standalone_site/src/features/Playlist/features/UpdatePlaylist/components/AddUserAccessButton.spec.tsx
@@ -1,0 +1,18 @@
+import { screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { render } from 'lib-tests';
+
+import { AddUserAccessButton } from './AddUserAccessButton';
+
+describe('AddUserAccessButton', () => {
+  it('opens AddUserAccessForm', () => {
+    render(<AddUserAccessButton playlistId="1" />);
+    expect(
+      screen.queryByRole('heading', { name: 'Add member' }),
+    ).not.toBeInTheDocument();
+    userEvent.click(screen.getByRole('button', { name: 'Add people' }));
+    expect(
+      screen.getByRole('heading', { name: 'Add member' }),
+    ).toBeInTheDocument();
+  });
+});